refactor: extract ALE launch dry-run CLI commands
